+++ packages/apps/reputation-oracle/server/src/modules/auth/types.ts
@@ -1,0 +1,4 @@
+export type AuthTokens = {
+  accessToken: string;
+  refreshToken: string;
+};
