+++ packages/apps/reputation-oracle/server/src/modules/kyc/types.ts
@@ -1,0 +1,7 @@
+export type VeriffCreateSessionResponse = {
+  status: string;
+  verification: {
+    id: string;
+    url: string;
+  };
+};
