+++ packages/apps/reputation-oracle/server/src/modules/qualification/types.ts
@@ -1,0 +1,6 @@
+export type Qualification = {
+  reference: string;
+  title: string;
+  description: string;
+  expiresAt?: string;
+};
