+++ packages/apps/reputation-oracle/server/src/modules/reputation/reputation.repository.ts
@@ -7,12 +7,7 @@ import { BaseRepository } from '../../database';
 
 import { ReputationEntityType, ReputationOrderBy } from './constants';
 import { ReputationEntity } from './reputation.entity';
-
-export type ExclusiveReputationCriteria = {
-  chainId: number;
-  address: string;
-  type: ReputationEntityType;
-};
+import type { ExclusiveReputationCriteria } from './types';
 
 @Injectable()
 export class ReputationRepository extends BaseRepository<ReputationEntity> {
