+++ packages/apps/reputation-oracle/server/src/modules/reputation/types.ts
@@ -1,0 +1,16 @@
+import { ChainId } from '@human-protocol/sdk';
+
+import { ReputationEntityType, ReputationLevel } from './constants';
+
+export type ReputationData = {
+  chainId: ChainId;
+  address: string;
+  level: ReputationLevel;
+  role: ReputationEntityType;
+};
+
+export type ExclusiveReputationCriteria = {
+  chainId: number;
+  address: string;
+  type: ReputationEntityType;
+};
