+++ packages/apps/reputation-oracle/server/src/modules/web3/index.ts
@@ -1,2 +1,3 @@
 export { Web3Module } from './web3.module';
-export { WalletWithProvider, Web3Service } from './web3.service';
+export { Web3Service } from './web3.service';
+export type { WalletWithProvider } from './types';
