+++ packages/apps/reputation-oracle/server/src/modules/web3/types.ts
@@ -1,0 +1,10 @@
+import { ChainId } from '@human-protocol/sdk';
+
+import type { Provider, Wallet } from 'ethers';
+
+export type Chain = {
+  id: ChainId;
+  rpcUrl: string;
+};
+
+export type WalletWithProvider = Wallet & { provider: Provider };
