+++ packages/apps/reputation-oracle/server/src/modules/web3/web3.service.spec.ts
@@ -1,9 +1,12 @@
 import { createMock } from '@golevelup/ts-jest';
+import { faker } from '@faker-js/faker';
 import { Test } from '@nestjs/testing';
 import { FeeData, JsonRpcProvider, Provider } from 'ethers';
-import { faker } from '@faker-js/faker';
-import { WalletWithProvider, Web3Service } from './web3.service';
+
 import { Web3ConfigService } from '../../config';
+
+import { Web3Service } from './web3.service';
+import type { WalletWithProvider } from './types';
 
 import { generateTestnetChainId, mockWeb3ConfigService } from './fixtures';
