+++ packages/apps/reputation-oracle/server/src/modules/web3/web3.service.ts
@@ -1,7 +1,8 @@
 import { ChainId } from '@human-protocol/sdk';
 import { Injectable } from '@nestjs/common';
-import { Provider, Wallet, ethers } from 'ethers';
+import { Wallet, ethers } from 'ethers';
 import { Web3ConfigService, Web3Network } from '../../config';
+import type { Chain, WalletWithProvider } from './types';
 
 export const supportedChainIdsByNetwork = {
   [Web3Network.MAINNET]: [ChainId.POLYGON, ChainId.BSC_MAINNET],
@@ -12,13 +13,6 @@ export const supportedChainIdsByNetwork = {
   ],
   [Web3Network.LOCAL]: [ChainId.LOCALHOST],
 } as const;
-
-type Chain = {
-  id: ChainId;
-  rpcUrl: string;
-};
-
-export type WalletWithProvider = Wallet & { provider: Provider };
 
 @Injectable()
 export class Web3Service {
